+++ scanpipe/apps.py
@@ -25,6 +25,9 @@ from pathlib import Path
 from django.apps import AppConfig
 from django.utils.translation import gettext_lazy as _
 
+from scanpipe.logging import RunLogger
+from scanpipe.logging import extra_logging
+
 
 class ScanPipeConfig(AppConfig):
     name = "scanpipe"
@@ -44,6 +47,12 @@ class ScanPipeConfig(AppConfig):
                 location = str(child.relative_to(project_root))
                 name = child.name[: -len(dot_py_suffix)]
                 self.pipelines.append((location, name))
+
+        # Decorates the default metaflow logger to capture log messages
+        # Warning: This import cannot be moved outside this method
+        from metaflow import cli
+
+        cli.logger = extra_logging(cli.logger, RunLogger())
 
     def is_valid(self, pipeline):
         """