+++ CHANGELOG.rst
@@ -3,6 +3,9 @@
 
 ### v1.0.7 (unreleased)
 
+- Log messages from Pipeline execution on a new Run instance `log` field
+  https://github.com/nexB/scancode.io/issues/66
+
 - Add support for scancode pipes and Project name with whitespaces
 
 - Add a profile() method on the Run model for profiling pipeline execution